TG-7: add messages in validations, added function in controllers, change model

+++ src/controllers/projects.js
@@ -1,18 +1,41 @@
 import Projects from '../models/Projects';
 
+const { ObjectId } = require('mongoose').Types;
+
+function isValidObjectId(id) {
+  if (ObjectId.isValid(id)) {
+    if ((String)(new ObjectId(id)) === id) { return true; }
+    return false;
+  }
+  return false;
+}
+
 const deleteProject = async (req, res) => {
   try {
     const { id } = req.params;
-    const result = await Projects.findByIdAndDelete(id);
-    return res.status(200).json({
-      message: `Project whit id ${id} deleted.`,
-      data: result,
-      error: false,
+    if (isValidObjectId(id)) {
+      const result = await Projects.findByIdAndDelete(id);
+      if (result !== null) {
+        return res.status(200).json({
+          message: `Project whit id ${id} deleted.`,
+          data: result,
+          error: false,
+        });
+      }
+      return res.status(404).json({
+        message: `there is not any project with that id ${id}.`,
+        error: true,
+      });
+    }
+    return res.status(400).json({
+      message: `Invalid id ${id}.`,
+      error: true,
     });
   } catch (error) {
-    return res.json({
-      message: 'An error dale papa',
-      error,
+    return res.status(500)({
+      message: error,
+      data: undefined,
+      error: true,
     });
   }
 };
@@ -25,16 +48,16 @@ const editProject = async (req, res) => {
       { ...req.body },
       { new: true },
     );
-
     return res.status(200).json({
-      message: `Project whit id ${id} edited`,
+      message: `Modified project whit id ${id}`,
       data: result,
       error: false,
     });
   } catch (error) {
-    return res.json({
-      message: 'An error occurred',
-      error,
+    return res.status(500)({
+      message: error,
+      data: undefined,
+      error: true,
     });
   }
 };
